[JSC] Unify SSACalculator https://bugs.webkit.org/show_bug.cgi?id=317635 rdar://180381307
Reviewed by Keith Miller.

Now B3::BasicBlock and DFG::BasicBlock has similar shape. So we can
easily share SSACalculator with them. This patch moves DFGSSACalculator
from JSC to WTF and use them from DFG and B3. This is following to the
similar manner in WTF::Liveness unification.

Since we moved all logic in B3::SSACalculator, it also brings path
compression to WTF::SSACalculator. Previously DFG::SSACalculator didn't
have that.

* Source/JavaScriptCore/JavaScriptCore.xcodeproj/project.pbxproj:
* Source/JavaScriptCore/Sources.txt:
* Source/JavaScriptCore/b3/B3SSACalculator.cpp: Removed.
* Source/JavaScriptCore/b3/B3SSACalculator.h:
(JSC::B3::SSACalculatorAdapter::SSACalculatorAdapter):
(JSC::B3::SSACalculatorAdapter::cfg const):
(JSC::B3::SSACalculatorAdapter::dominators):
(JSC::B3::SSACalculatorAdapter::dumpValue):
(JSC::B3::SSACalculator::Variable::index const): Deleted.
(JSC::B3::SSACalculator::Variable::Variable): Deleted.
(JSC::B3::SSACalculator::Def::variable const): Deleted.
(JSC::B3::SSACalculator::Def::block const): Deleted.
(JSC::B3::SSACalculator::Def::value const): Deleted.
(JSC::B3::SSACalculator::Def::Def): Deleted.
(JSC::B3::SSACalculator::variable): Deleted.
(JSC::B3::SSACalculator::computePhis): Deleted.
(): Deleted.
(JSC::B3::SSACalculator::reachingDefAtHead): Deleted.
* Source/JavaScriptCore/dfg/DFGSSACalculator.cpp: Removed.
* Source/JavaScriptCore/dfg/DFGSSACalculator.h:
(JSC::DFG::SSACalculatorAdapter::SSACalculatorAdapter):
(JSC::DFG::SSACalculatorAdapter::cfg const):
(JSC::DFG::SSACalculatorAdapter::dominators):
(JSC::DFG::SSACalculatorAdapter::dumpValue):
(JSC::DFG::SSACalculator::Variable::index const): Deleted.
(JSC::DFG::SSACalculator::Variable::Variable): Deleted.
(JSC::DFG::SSACalculator::Def::variable const): Deleted.
(JSC::DFG::SSACalculator::Def::block const): Deleted.
(JSC::DFG::SSACalculator::Def::value const): Deleted.
(JSC::DFG::SSACalculator::Def::Def): Deleted.
(JSC::DFG::SSACalculator::variable): Deleted.
(JSC::DFG::SSACalculator::computePhis): Deleted.
(): Deleted.
(JSC::DFG::SSACalculator::reachingDefAtHead): Deleted.
* Source/WTF/WTF.xcodeproj/project.pbxproj:
* Source/WTF/wtf/CMakeLists.txt:
* Source/WTF/wtf/SSACalculator.h: Copied from Source/JavaScriptCore/dfg/DFGSSACalculator.h.
(WTF::SSACalculator::SSACalculator):
(WTF::SSACalculator::reset):
(WTF::SSACalculator::Variable::index const):
(WTF::SSACalculator::Variable::dump const):
(WTF::SSACalculator::Variable::dumpVerbose const):
(WTF::SSACalculator::Variable::Variable):
(WTF::SSACalculator::Def::variable const):
(WTF::SSACalculator::Def::block const):
(WTF::SSACalculator::Def::value const):
(WTF::SSACalculator::Def::dump const):
(WTF::SSACalculator::Def::Def):
(WTF::SSACalculator::newVariable):
(WTF::SSACalculator::newDef):
(WTF::SSACalculator::variable):
(WTF::SSACalculator::computePhis):
(WTF::SSACalculator::nonLocalReachingDef):
(WTF::SSACalculator::reachingDefAtHead):
(WTF::SSACalculator::reachingDefAtTail):
(WTF::SSACalculator::dump const):
(WTF::SSACalculator::ensureDominanceFrontiers):

Canonical link: https://commits.webkit.org/315756@main
