+++ Source/JavaScriptCore/Sources.txt
@@ -171,7 +171,6 @@ b3/B3Procedure.cpp
 b3/B3PureCSE.cpp
 b3/B3ReduceDoubleToFloat.cpp
 b3/B3ReduceStrength.cpp
-b3/B3SSACalculator.cpp
 b3/B3SIMDValue.cpp
 b3/B3SlotBaseValue.cpp
 b3/B3StackmapGenerationParams.cpp
@@ -434,7 +433,6 @@ dfg/DFGPromotedHeapLocation.cpp
 dfg/DFGPureValue.cpp
 dfg/DFGPutStackSinkingPhase.cpp
 dfg/DFGRegisteredStructureSet.cpp
-dfg/DFGSSACalculator.cpp
 dfg/DFGSSAConversionPhase.cpp
 dfg/DFGSSALoweringPhase.cpp
 dfg/DFGSnippetParams.cpp
