--- Source/JavaScriptCore/b3/B3SSACalculator.cpp
@@ -1,179 +1,0 @@
-/*
- * Copyright (C) 2016 Apple Inc. All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions
- * are met:
- * 1. Redistributions of source code must retain the above copyright
- *    notice, this list of conditions and the following disclaimer.
- * 2. Redistributions in binary form must reproduce the above copyright
- *    notice, this list of conditions and the following disclaimer in the
- *    documentation and/or other materials provided with the distribution.
- *
- * THIS SOFTWARE IS PROVIDED BY APPLE INC. ``AS IS'' AND ANY
- * EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
- * PURPOSE ARE DISCLAIMED.  IN NO EVENT SHALL APPLE INC. OR
- * CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
- * EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
- * PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
- * PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
- * OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE. 
- */
-
-#include "config.h"
-#include "B3SSACalculator.h"
-
-#if ENABLE(B3_JIT)
-
-#include <wtf/CommaPrinter.h>
-#include <wtf/TZoneMallocInlines.h>
-
-namespace JSC { namespace B3 {
-
-WTF_MAKE_TZONE_ALLOCATED_IMPL(SSACalculator);
-
-void SSACalculator::Variable::dump(PrintStream& out) const
-{
-    out.print("var", m_index);
-}
-
-void SSACalculator::Variable::dumpVerbose(PrintStream& out) const
-{
-    dump(out);
-    if (!m_blocksWithDefs.isEmpty()) {
-        out.print("(defs: "_s);
-        CommaPrinter comma;
-        for (BasicBlock* block : m_blocksWithDefs)
-            out.print(comma, *block);
-        out.print(")"_s);
-    }
-}
-
-void SSACalculator::Def::dump(PrintStream& out) const
-{
-    out.print("def(", *m_variable, ", ", *m_block, ", ", pointerDump(m_value), ")");
-}
-
-SSACalculator::SSACalculator(Procedure& proc)
-    : m_data(proc.size())
-    , m_proc(proc)
-{
-}
-
-SSACalculator::~SSACalculator() = default;
-
-void SSACalculator::ensureDominanceFrontiers()
-{
-    if (m_dominanceFrontiersComputed)
-        return;
-    m_dominanceFrontiersComputed = true;
-
-    m_dominanceFrontiers = IndexMap<BasicBlock*, Vector<BasicBlock*>>(m_proc.size());
-
-    // "A Simple, Fast Dominance Algorithm", Cooper, Harvey, and Kennedy, 2001.
-    // https://www.cs.tufts.edu/~nr/cs257/archive/keith-cooper/dom14.pdf
-    for (BasicBlock* block : m_proc) {
-        if (block->numPredecessors() < 2)
-            continue;
-        BasicBlock* idom = m_dominators->idom(block);
-        for (BasicBlock* predecessor : block->predecessors()) {
-            BasicBlock* runner = predecessor;
-            while (runner && runner != idom) {
-                auto& frontier = m_dominanceFrontiers[runner];
-                if (frontier.isEmpty() || frontier.last() != block)
-                    frontier.append(block);
-                runner = m_dominators->idom(runner);
-            }
-        }
-    }
-}
-
-void SSACalculator::reset()
-{
-    m_variables.clear();
-    m_defs.clear();
-    m_phis.clear();
-    for (unsigned blockIndex = m_data.size(); blockIndex--;) {
-        m_data[blockIndex].m_defs.clear();
-        m_data[blockIndex].m_phis.clear();
-    }
-    m_dominanceFrontiersComputed = false;
-}
-
-SSACalculator::Variable* SSACalculator::newVariable()
-{
-    return &m_variables.alloc(Variable(m_variables.size()));
-}
-
-SSACalculator::Def* SSACalculator::newDef(Variable* variable, BasicBlock* block, Value* value)
-{
-    Def* def = m_defs.add(Def(variable, block, value));
-    auto result = m_data[block].m_defs.add(variable, def);
-    if (result.isNewEntry)
-        variable->m_blocksWithDefs.append(block);
-    else
-        result.iterator->value = def;
-    return def;
-}
-
-SSACalculator::Def* SSACalculator::nonLocalReachingDef(BasicBlock* block, Variable* variable)
-{
-    return reachingDefAtTail(m_dominators->idom(block), variable);
-}
-
-SSACalculator::Def* SSACalculator::reachingDefAtTail(BasicBlock* startingBlock, Variable* variable)
-{
-    for (BasicBlock* block = startingBlock; block; block = m_dominators->idom(block)) {
-        if (Def* def = m_data[block].m_defs.get(variable)) {
-            for (BasicBlock* otherBlock = startingBlock; otherBlock != block; otherBlock = m_dominators->idom(otherBlock))
-                m_data[otherBlock].m_defs.add(variable, def);
-            return def;
-        }
-    }
-    return nullptr;
-}
-
-void SSACalculator::dump(PrintStream& out) const
-{
-    out.print("<Variables: ["_s);
-    CommaPrinter comma;
-    for (unsigned i = 0; i < m_variables.size(); ++i) {
-        out.print(comma);
-        m_variables[i].dumpVerbose(out);
-    }
-    out.print("], Defs: ["_s);
-    comma = CommaPrinter();
-    for (Def* def : const_cast<SSACalculator*>(this)->m_defs)
-        out.print(comma, *def);
-    out.print("], Phis: ["_s);
-    comma = CommaPrinter();
-    for (Def* def : const_cast<SSACalculator*>(this)->m_phis)
-        out.print(comma, *def);
-    out.print("], Block data: ["_s);
-    comma = CommaPrinter();
-    for (unsigned blockIndex = 0; blockIndex < m_proc.size(); ++blockIndex) {
-        BasicBlock* block = m_proc[blockIndex];
-        if (!block)
-            continue;
-        
-        out.print(comma, *block, "=>("_s);
-        out.print("Defs: {"_s);
-        CommaPrinter innerComma;
-        for (auto entry : m_data[block].m_defs)
-            out.print(innerComma, *entry.key, "->"_s, *entry.value);
-        out.print("}, Phis: {"_s);
-        innerComma = CommaPrinter();
-        for (Def* def : m_data[block].m_phis)
-            out.print(innerComma, *def);
-        out.print("})"_s);
-    }
-    out.print("]>"_s);
-}
-
-} } // namespace JSC::B3
-
-#endif // ENABLE(B3_JIT)
-
