+++ Source/WTF/wtf/CMakeLists.txt
@@ -297,6 +297,7 @@ set(WTF_PUBLIC_HEADERS
     SHA1.h
     SIMDHelpers.h
     SIMDUTF.h
+    SSACalculator.h
     SafeStrerror.h
     SaturatedArithmetic.h
     SchedulePair.h
